# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MEX API",
    "version": "1.0.0",
    "description": "API documentation of medical examination system (T5)"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/api/file/{id}": {
      "post": {
        "tags": ["Examination pdf file"],
        "summary": "Upload a file and update examination record",
        "description": "Uploads a file to S3 and updates the `examination_filename` field in the database for a specific MEX record.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the MEX record to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "File is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "File blob is required."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error while uploading file.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error uploading file"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/file/{mex_id}": {
      "get": {
        "tags": ["Examination pdf file"],
        "summary": "Retrieve file URLs for a specific MEX record file",
        "description": "Fetches the examination file associated with a given MEX ID and returns a pre-signed URL for downloading.",
        "parameters": [
          {
            "in": "path",
            "name": "mex_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the MEX record whose file URLs need to be retrieved."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved file URLs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "fileUrl": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://s3.example.com/bucket/file.pdf"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error loading image.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error loading image"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Examination pdf file"],
        "summary": "Delete an examination file for a specific MEX record",
        "description": "Deletes the examination file associated with a given MEX ID from storage and updates the record.",
        "parameters": [
          {
            "in": "path",
            "name": "mex_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the MEX record whose file needs to be deleted."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the file.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "file deleted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error while deleting the file.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "error while delete"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}